+++ KeireCore/Include/KeireInternal/Scripting/ManagedRuntimeBindings.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "KeireInternal/Scripting/ManagedRuntimeFoundation.h"
+#include "KeireInternal/Scripting/ManagedRuntimePhysics.h"
 #include "KeireInternal/Scripting/ManagedRuntimeRendering.h"
 #include "KeireInternal/Scripting/ManagedRuntimeUi.h"
 #include "KeireInternal/Scripting/ManagedRuntimeWorld.h"
@@ -17,6 +18,7 @@ namespace Keire::Detail
 
       private:
         ManagedRuntimeFoundationScope m_Foundation;
+        ManagedRuntimePhysicsScope m_Physics;
         ManagedRuntimeWorldScope m_World;
         ManagedRuntimeRenderingScope m_Rendering;
         ManagedRuntimeUiScope m_Ui;